Migrate loading/loading to TypeScript
